llama index/requirements.txt
llama-index
openai
pypdf
python-dotenv
fastapi
streamlit